# Fix `Quad.Contains` sometimes failing due to floating point precision issues