--- a/osu.Framework/Graphics/Primitives/Quad.cs
+++ b/osu.Framework/Graphics/Primitives/Quad.cs
@@ -108,8 +108,10 @@ namespace osu.Framework.Graphics.Primitives
         public ReadOnlySpan<Vector2> GetVertices() => MemoryMarshal.CreateReadOnlySpan(ref Unsafe.AsRef(in TopLeft), 4);
 
         public bool Contains(Vector2 pos) =>
-            new Triangle(BottomRight, BottomLeft, TopRight).Contains(pos) ||
-            new Triangle(TopLeft, TopRight, BottomLeft).Contains(pos);
+            Vector2.PerpDot(BottomLeft - TopLeft, pos - TopLeft) <= 0
+            && Vector2.PerpDot(BottomRight - BottomLeft, pos - BottomLeft) <= 0
+            && Vector2.PerpDot(TopRight - BottomRight, pos - BottomRight) <= 0
+            && Vector2.PerpDot(TopLeft - TopRight, pos - TopRight) <= 0;
 
         /// <summary>
         /// Computes the area of this <see cref="Quad"/>.