Closes #5669.

I will pre-empt this by saying that I am only *mostly* confident of this fix, so please be on the look-out if you don't like anything here. It took me the better part of an hour to convince myself that this works due to reasons I will elaborate on below.

To the point: In order to avoid the issue demonstrated by the test in #5668, a different way of computing whether a given point is inside of the quad is used. Namely, this method uses the perpendicular dot product, which has the property of

$$ \textbf{a}^\perp \cdot \textbf{b} = |\textbf{a}| |\textbf{b}| \sin \theta $$

wherein $\theta$ is the angle between the vectors $\textbf{a}$ and $\textbf{b}$.[^1] In particular, this means that because the sine of an angle is positive in the range [0, 180deg], and negative in the range [-180deg, 0], it allows to distinguish which side of the half-plane defined by the direction of vector $\textbf{a}$ the vector $\textbf{b}$ lies. On the Cartesian plane, it will be positive if $\textbf{b}$ is "on the left of" or "counterclockwise" to $\textbf{a}$, zero if they are collinear, and negative otherwise.

Assuming `Quad` is convex (which can be done, as it inherits `IConvexPolygon` - and notably, the previous implementation also silently assumed convexity, because in the general case you can't just arbitrarily slice the quad into two tris if it's not convex), we can say that a point lies within the quad if it lies on the same side of four half-planes defined by the quad's sides (by which I mean, for every side, we can divide the plane using a line going through both ends of the segment, and take the half-plane where the quad's interior is). Due to this we can therefore check whether for every vertex of the quad, the sign of the perp dot product between the vector which starts at the current vertex and ends at the next vertex (going in counterclockwise order), and the vector which starts at the current vertex and ends at the point being tested by `Contains()`, is the same - if it is, then the tested point is inside of the quad.

Now this is the part where this gets confusing: On the Cartesian plane, the sign we want would be *positive* (or zero, as zero means collinearity). But because computer graphics are hell and somebody somewhere decided that when drawing to the screen Y should be *down* and not up and that is what `Quad` abides by, the signs need to flip. So for our test here we use zero or less.

Interestingly, I have managed to produce a benchmark (included in the diff) which happens to show that this method is faster:

### Before

|   Method | NumPoints |        Mean |     Error |    StdDev | Allocated |
|--------- |---------- |------------:|----------:|----------:|----------:|
| Contains |        10 |    274.5 ns |   1.10 ns |   1.03 ns |         - |
| Contains |       100 |  2,903.4 ns |   3.65 ns |   3.41 ns |         - |
| Contains |      1000 | 29,330.9 ns | 297.09 ns | 263.36 ns |         - |

### After

|   Method | NumPoints |        Mean |     Error |    StdDev | Allocated |
|--------- |---------- |------------:|----------:|----------:|----------:|
| Contains |        10 |    62.32 ns |  0.726 ns |  0.679 ns |         - |
| Contains |       100 |   573.51 ns |  3.283 ns |  2.742 ns |         - |
| Contains |      1000 | 5,937.07 ns | 62.791 ns | 52.433 ns |         - |

but this is very likely highly dependent on data. The new method will be likely faster for rejections (as checking half-planes may allow to reject a point faster) but slower for hits (as all four half-planes must be examined), which is probably the correct access pattern to optimise for anyways?

As for testing, I did some using the draw visualiser but admittedly not much. I have checked that this diff passes game-side tests which is something extra I suppose?

[^1]: https://mathworld.wolfram.com/PerpDotProduct.html