@peppy 2023-03-08
[Blaming back](https://github.com/ppy/osu-framework/blame/8c69bbf03e1ccbcee508a7ea709c3391ddaa8ab7/osu.Framework/Graphics/Primitives/Quad.cs#L114C25-L121
) to try and find the original code owner ends at the initial framework commit 😅.

@Tom94 I don't remember if you were behind the original implementation of this (or if it was taken from another project), but this might be something you're interested in taking a quick look at.

The math looks to check out as far as I can tell, and fixes the case I was encountering 👍 .