@Tom94 2023-03-08
Ha, yes, I remember hacking the `Contains` function using two `Triangles`. It already felt hacky back then and the new implementation is indeed better in every respect as long as we don't create non-convex quads anywhere. (We shouldn't, if all we do is trainsform rectangles with matrices.)